LPS-105380 Move class
